+++ web/.storybook/main.ts
@@ -10,6 +10,7 @@ import {
 
 const STORY_EXTENSIONS = "@(js|jsx|mjs|ts|tsx)";
 const DESIGN_COMPONENT_STORIES = [
+  "Codeblock/Codeblock",
   "LangfuseIcon/LangfuseIcon",
   "Progress/Progress",
   "Spinner/Spinner",
