+++ web/src/components/ui/MarkdownViewer.tsx
@@ -11,7 +11,7 @@ import {
 } from "react";
 import ReactMarkdown, { type Options } from "react-markdown";
 import remarkGfm from "remark-gfm";
-import { CodeBlock } from "@/src/components/ui/Codeblock";
+import { CodeBlock } from "@/src/components/design-system/Codeblock/Codeblock";
 import { useTheme } from "next-themes";
 import { ImageOff, Info } from "lucide-react";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
